requirements.txt
django-cors-headers
pymysql
djangorestframework
django-environ